Renamed balldetector for consistency
